backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
neo4j>=4.4.0
python-dotenv>=0.19.0
requests>=2.25.1
openai>=0.27.0
groq>=0.4.0
faiss-cpu>=1.7.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
faster-whisper>=0.5.0
openmeteo-requests>=1.0.0
openmeteo-sdk>=1.0.0